requirements.txt
Django==1.11.7
django-debug-toolbar==1.8
django-widget-tweaks==1.4.1
Markdown==2.6.9
django-oauth-toolkit==1.0.0
django-cors-middleware==1.3.1
djangorestframework==3.8.2
Pygments==2.2.0